+++ rust/src/regex_stress_test.rs
@@ -139,7 +139,8 @@ mod regex_stress_tests {
             
             // TODO: Uncomment when actual parser is available
             /*
-            let mut parser = RegexParser::with_debug(test_input);
+            let test_name = format!("stress_test_{}", i + 1);
+            let mut parser = RegexParser::with_debug_log(test_input, &test_name);
             let parse_start = Instant::now();
             
             match parser.parse() {